requirements.txt
openai
moviepy
Pillow
discord.py